API: Move some enums from `marker_api::ast` -> `marker_api::common`
