+++ marker_api/src/ast/expr/block_expr.rs
@@ -1,6 +1,6 @@
 use crate::{
-    ast::{pat::PatKind, stmt::StmtKind, ty::SynTyKind, Safety, Syncness},
-    common::{BodyId, SpanId},
+    ast::{pat::PatKind, stmt::StmtKind, ty::SynTyKind},
+    common::{BodyId, Safety, SpanId, Syncness},
     context::with_cx,
     ffi::{FfiOption, FfiSlice},
     span::{Ident, Span},
