+++ marker_api/src/ast/item/extern_block_item.rs
@@ -1,5 +1,4 @@
-use crate::ast::Abi;
-use crate::ffi::FfiSlice;
+use crate::{common::Abi, ffi::FfiSlice};
 
 use super::{CommonItemData, ExternItemKind};
